improved layout of new meal/ edit meal form, changed background throughout application, made styles consistent throughout the app
