+++ src/app/app-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { CreateOrUpdateMealFormComponent } from './components/create-or-update-meal-form/create-or-update-meal-form.component';
 import { ErrorPageComponent } from './components/error-page/error-page.component';
 import { LogInPageComponent } from './components/log-in-page/log-in-page.component';
 import { MainPageComponent } from './components/main-page/main-page.component';
 import { MealListComponent } from './components/meal-list/meal-list.component';
 import { MealViewComponent } from './components/meal-view/meal-view.component';
-import { NewMealComponent } from './components/new-meal/new-meal.component';
 
 const routes: Routes = [{
   path: '', component: MainPageComponent
@@ -17,10 +17,10 @@ const routes: Routes = [{
   path: 'log-in-page', component: LogInPageComponent
 },
 {
-  path: 'meal-list/new-meal', component: NewMealComponent
+  path: 'meal/new-meal', component: CreateOrUpdateMealFormComponent
 },
 {
-  path: 'meal-list/meal-id', component: MealViewComponent
+  path: 'meal/:id/edit', component: CreateOrUpdateMealFormComponent
 },
 {
   path: '**', component: ErrorPageComponent
