+++ src/app/components/create-or-update-meal-form/create-or-update-meal-form.component.html
@@ -1,0 +1,107 @@
+<div class="form-container"
+     [formGroup]="form">
+
+    <div class="name-container">
+        <h3>{{ id ? 'Edytuj obiad' : 'Dodaj nowy obiad' }}</h3>
+
+        <mat-form-field>
+            <mat-label>Nazwa obiadu</mat-label>
+            <input matInput
+                   formControlName="name"
+                   maxlength="50">
+            <mat-hint align="end">{{name.value.length}} / 50</mat-hint>
+
+        </mat-form-field>
+        <div *ngIf="name.invalid && (name.dirty || name.touched)"
+             class="validation-container">
+            <div *ngIf="name.errors.required">
+                Pole wymagane
+            </div>
+            <div *ngIf="name.errors.minlength">
+                Pole powinno zawierać od 3 do 50 liter.
+            </div>
+            <div *ngIf="name.errors.pattern">
+                Pole zawiera niedopuszczalne znaki. Pole powinno zawierać tylko litery.
+            </div>
+        </div>
+    </div>
+
+    <div class="image-container">
+        <h3>{{ id ? 'Zdjęcie' : 'Zdjęcie (opcjonalne)'}}</h3>
+        <div class="dropzone"
+             appDragAndDrop
+             *ngIf="!image"
+             (fileDrop)="onFileDropped($event)">
+            <span class="material-icons drop-icon">
+                image_search
+            </span>
+            <h4>Przeciągnij zdjęcie</h4>
+            <h4>lub</h4>
+            <label for="upload-img">Przeglądaj...</label>
+            <input type="file"
+                   name="image"
+                   id="upload-img"
+                   accept="image/*"
+                   formControlName="imageUrl"
+                   (change)="fileBrowseHandler($event)">
+        </div>
+        <div class="inserted-img-container"
+             *ngIf="image">
+            <button class="delete-image-btn"
+                    mat-mini-fab
+                    (click)="deleteImage()">
+                <mat-icon>close</mat-icon>
+            </button>
+            <img class="inserted-img"
+                 [src]="image.url">
+        </div>
+    </div>
+
+    <div class="list-container">
+        <h3>{{ id ? 'Lista składników' : 'Lista składników (opcjonalne)'}}</h3>
+        <div cdkDropList
+             class="ingredients-list"
+             (cdkDropListDropped)="dropIngredients($event)">
+            <div class="ingredient-box"
+                 *ngFor="let ingredient of ingredientsList; index as i"
+                 cdkDrag>
+                {{ingredient}}
+                <button class="delete-item-list-btn"
+                        mat-mini-fab
+                        (click)="deleteIngredient(i)">
+                    <mat-icon>close</mat-icon>
+                </button>
+            </div>
+            <div>
+            </div>
+        </div>
+        <button class="add-item-list-btn"
+                mat-mini-fab
+                (click)="openIngredientDialog()">
+            <mat-icon>add</mat-icon>
+        </button>
+    </div>
+
+
+    <div class="recipe-container">
+        <h3>{{ id ? 'Przepis' : 'Przepis (opcjonalne)'}}</h3>
+        <quill-editor class="recipe-area"
+                      formControlName="recipe"
+                      placeholder=""></quill-editor>
+    </div>
+
+    <div class="btn-container">
+        <button routerLink="/meal-list"
+                class="back-btn">
+            Wróć do listy
+        </button>
+        <button class="add-meal-btn"
+                [disabled]="form.invalid"
+                [ngStyle]="{'background-color': form.valid ? '#1b84b4' : 'grey'}"
+                (click)="saveMeal()"
+                routerLink="/meal-list">
+            Zapisz obiad
+        </button>
+
+    </div>
+</div>
