+++ src/app/components/log-in-form/log-in-form.component.html
@@ -12,6 +12,6 @@
                      formControlName="password"
                      placeholder="Hasło">
        </div>
-       <button>ZALOGUJ</button>
+       <button>Zaloguj</button>
        <button class="new-password-button">Zapomniałeś hasła?</button>
 </div>
