+++ src/app/components/log-in-page/log-in-page.component.scss
@@ -1,5 +1,9 @@
 .container {
     display: flex;
     justify-content: center;
-    margin-top: 50px;
+    width: fit-content;
+    margin: 50px auto;
+    padding: 30px;
+    background-color: #ffffff;
+    border-radius: 2px;
 }
