+++ src/app/components/main-page/main-page.component.html
@@ -1,6 +1,6 @@
 <div class="page">
-    <h1>Hej imię! Jak się masz?</h1>
-    <h2>Ostatnio wylosowane obiady</h2>
+    <h3>Hej imię! Jak się masz?</h3>
+    <h3>Ostatnio wylosowane obiady</h3>
     <div class="last-meals-container">
         <app-last-drawed-meals></app-last-drawed-meals>
     </div>
@@ -9,7 +9,7 @@
         <button class="btn"
                 routerLink="meal-list">Przejdź do listy obiadów</button>
     </div>
-    <h2>Plan tygodnia</h2>
+    <h3>Plan tygodnia</h3>
     <table>
         <tr>
             <th>Poniedziałek</th>
@@ -26,7 +26,5 @@
             <td *ngFor="let meal of drawedMeals"
                 cdkDrag>{{meal}}</td>
         </tr>
-
     </table>
-
 </div>
