+++ src/app/components/main-page/main-page.component.scss
@@ -1,15 +1,15 @@
 .page {
-    background-color: #dad0d0;
-    height: 100%;
-    padding-left: 20px;
-    padding-top: 20px;
+    background-color: #ffffff;
+    border-radius: 2px;
+    margin: 50px auto;
+    width: min-content;
+    padding: 30px;
 }
 
-h1,
-h2 {
-    color: #8b8888;
+h3 {
+    color: #38424c;
+    font-size: 22px;
     text-align: center;
-    margin-top: 20px;
 }
 
 .last-meals-container {
@@ -28,6 +28,7 @@ h2 {
     margin-bottom: 5px;
     border: none;
     background-color: #ffffff;
+    border: solid 1px #8b8888;
     font-size: 10px;
     font-weight: bold;
     border-radius: 2px;
@@ -36,6 +37,7 @@ h2 {
 
 .blue-btn {
     background-color: #1b84b4;
+    border: none;
     color: #ffffff;
 }
