+++ src/app/components/meal-list/meal-list.component.html
@@ -1,43 +1,45 @@
-<div class="container">
-    <div class="display-options-container">
-        <button class="display-btn"
-                (click)="changeLayout('gallery')">
-            <mat-icon class="display-option">
-                grid_view
-            </mat-icon>
-        </button>
-        <button class="display-btn"
-                (click)="changeLayout('list')">
-            <mat-icon class="display-option">
-                format_list_bulleted
-            </mat-icon>
+<div class="page">
+    <div class="container">
+        <div class="display-options-container">
+            <button class="display-btn"
+                    (click)="changeLayout('gallery')">
+                <mat-icon class="display-option">
+                    grid_view
+                </mat-icon>
+            </button>
+            <button class="display-btn"
+                    (click)="changeLayout('list')">
+                <mat-icon class="display-option">
+                    format_list_bulleted
+                </mat-icon>
+            </button>
+        </div>
+        <div class="sort-options-container">
+            <mat-form-field appearance="fill">
+                <mat-label>Sortuj</mat-label>
+                <mat-select [formControl]="sortByOptionControl">
+                    <mat-option *ngFor="let dropdownOption of sortByDropdownValues"
+                                [value]="dropdownOption.sortByOption">
+                        {{dropdownOption.displayValue}}
+                    </mat-option>
+                </mat-select>
+            </mat-form-field>
+        </div>
+    </div>
+    <div class="add-btn-container">
+        <h3>Dodaj nowy obiad</h3>
+        <button class="add-btn"
+                mat-mini-fab
+                routerLink="/meal/new-meal">
+            <mat-icon>add</mat-icon>
         </button>
     </div>
-    <div class="sort-options-container">
-        <mat-form-field appearance="fill">
-            <mat-label>Sortuj</mat-label>
-            <mat-select [formControl]="sortByOptionControl">
-                <mat-option *ngFor="let dropdownOption of sortByDropdownValues"
-                            [value]="dropdownOption.sortByOption">
-                    {{dropdownOption.displayValue}}
-                </mat-option>
-            </mat-select>
-        </mat-form-field>
-    </div>
-</div>
-<div class="add-btn-container">
-    <h3>Dodaj nowy obiad</h3>
-    <button class="add-btn"
-            mat-mini-fab
-            routerLink="new-meal">
-        <mat-icon>add</mat-icon>
-    </button>
-</div>
-<div [ngClass]="{
+    <div [ngClass]="{
     'gallery-container':layoutType == 'gallery',
     'list-container':layoutType == 'list'}">
-    <app-meal *ngFor="let meal of meals | async"
-              [layoutType]="layoutType"
-              [meal]="meal"></app-meal>
+        <app-meal *ngFor="let meal of meals | async"
+                  [layoutType]="layoutType"
+                  [meal]="meal"></app-meal>
+    </div>
+    <h3>Liczba obiadów: {{ (meals | async).length }}</h3>
 </div>
-<h3>Liczba obiadów: {{ (meals | async).length }}</h3>
