+++ src/app/components/meal/meal.component.html
@@ -4,24 +4,31 @@
             <img [src]="meal.imageUrl">
             <button class="edit-btn"
                     mat-mini-fab
-                    routerLink="meal-id">
+                    [routerLink]="['/meal', meal.id, 'edit']">
                 <mat-icon>create</mat-icon>
             </button>
         </div>
         <div class="content">
             <h4>{{ meal.name }}</h4>
-            <h5 *ngIf="meal.lastDrawDate">Ostatnio wylosowany: {{ meal.lastDrawDate | date: "longDate" }}</h5>
-            <h5 *ngIf="!meal.lastDrawDate">Jeszcze nie wylosowano tego obiadu</h5>
+            <h5 *ngIf="meal.lastDrawDate; else elseMessage">
+                Ostatnio wylosowany: {{ meal.lastDrawDate | date: "longDate" }}
+            </h5>
+            <ng-template #elseMessage>
+                <h5>Jeszcze nie wylosowano tego obiadu</h5>
+            </ng-template>
+
         </div>
-        <mat-card-actions>
-            <button mat-button>SHARE</button>
-        </mat-card-actions>
+        <button class="add-btn">Dodaj do planu tygodnia</button>
     </mat-card>
 </div>
 <div *ngIf="layoutType == 'list'">
     <mat-card class="list-meal-card">
         <span>{{ meal.name }}</span>
-        <span *ngIf="meal.lastDrawDate">Ostatnio wylosowany: {{ meal.lastDrawDate | date: "longDate" }}</span>
-        <span *ngIf="!meal.lastDrawDate">Jeszcze nie wylosowano tego obiadu</span>
+        <span *ngIf="meal.lastDrawDate; else elseMessage">
+            Ostatnio wylosowany: {{ meal.lastDrawDate | date: "longDate" }}
+        </span>
+        <ng-template #elseMessage>
+            <span>Jeszcze nie wylosowano tego obiadu</span>
+        </ng-template>
     </mat-card>
 </div>
