+++ src/app/components/meal/meal.component.scss
@@ -24,6 +24,30 @@ img {
     top: 0px;
 }
 
+.add-btn {
+    width: 300px;
+    padding: 15px;
+    margin-bottom: 5px;
+    border: none;
+    font-size: 10px;
+    font-weight: bold;
+    border-radius: 2px;
+    cursor: pointer;
+    background-color: #1b84b4;
+    color: #ffffff;
+}
+
+h4 {
+    overflow: hidden;
+    text-overflow: ellipsis;
+    color: #38424c;
+    margin-top: 10px;
+}
+
+h5 {
+    color: #38424c;
+}
+
 .content {
     text-align: center;
     cursor: default;
