--- src/app/components/new-meal/new-meal.component.html
@@ -1,104 +1,0 @@
-<div class="btn-container">
-    <button routerLink="/meal-list"
-            class="back-btn">
-        Wróć do listy
-    </button>
-    <button class="add-meal-btn"
-            [disabled]="form.invalid"
-            [ngStyle]="{
-                'background-color': form.valid ? '#34bb19' : 'grey'
-            }"
-            (click)="saveMeal()"
-            routerLink="/meal-list">
-        Dodaj obiad
-    </button>
-</div>
-<div class="form-container"
-     [formGroup]="form">
-    <div class="child-container">
-
-        <h3>Nowy obiad</h3>
-
-        <mat-form-field>
-            <mat-label>Nazwa obiadu</mat-label>
-            <input matInput
-                   formControlName="name"
-                   maxlength="50">
-            <mat-hint align="end">{{name.value.length}} / 50</mat-hint>
-
-        </mat-form-field>
-        <div *ngIf="name.invalid && (name.dirty || name.touched)"
-             class="validation-container">
-            <div *ngIf="name.errors.required">
-                Pole wymagane
-            </div>
-            <div *ngIf="name.errors.minlength">
-                Pole powinno zawierać od 3 do 50 liter.
-            </div>
-            <div *ngIf="name.errors.pattern">
-                Pole zawiera niedopuszczalne znaki. Pole powinno zawierać tylko litery.
-            </div>
-        </div>
-
-        <h3>Lista składników</h3>
-        <div class="list-container">
-            <div cdkDropList
-                 class="ingredients-list"
-                 (cdkDropListDropped)="dropIngredients($event)">
-                <div class="ingredient-box"
-                     *ngFor="let ingredient of ingredientsList; let i = index"
-                     cdkDrag>
-                    {{ingredient}}
-                    <button class="delete-item-list-btn"
-                            mat-mini-fab
-                            (click)="deleteIngredient(i)">
-                        <mat-icon>close</mat-icon>
-                    </button>
-                </div>
-                <div>
-                </div>
-            </div>
-            <button class="add-item-list-btn"
-                    mat-mini-fab
-                    (click)="openIngredientDialog()">
-                <mat-icon>add</mat-icon>
-            </button>
-        </div>
-
-    </div>
-    <div class="child-container">
-
-        <h3>Zdjęcie</h3>
-        <div class="dropzone"
-             appDragAndDrop
-             *ngIf="!image"
-             (fileDrop)="onFileDropped($event)">
-            <span class="material-icons drop-icon">
-                image_search
-            </span>
-            <h4>Drag and drop image</h4>
-            <h4>or</h4>
-            <label for="upload-img">Browse...</label>
-            <input type="file"
-                   name="image"
-                   id="upload-img"
-                   accept="image/*"
-                   formControlName="imageUrl"
-                   (change)="fileBrowseHandler($event)">
-        </div>
-        <div class="inserted-img-container"
-             *ngIf="image">
-            <button class="delete-image-btn"
-                    mat-mini-fab
-                    (click)="deleteImage()">
-                <mat-icon>close</mat-icon>
-            </button>
-            <img class="inserted-img"
-                 [src]="image.url">
-        </div>
-
-        <h3>Przepis</h3>
-        <quill-editor class="recipe-area"
-                      formControlName="recipe"></quill-editor>
-    </div>
-</div>
