--- src/app/components/new-meal/new-meal.component.scss
@@ -1,165 +1,0 @@
-.btn-container {
-    display: flex;
-    justify-content: space-between;
-    margin: 10px 30px;
-}
-
-button {
-    cursor: pointer;
-}
-
-.back-btn,
-.add-meal-btn {
-    width: 150px;
-    height: 50px;
-    border: none;
-    border-radius: 2px;
-    color: #ffffff;
-    background-color: #ec8a3a;
-    font-weight: bold;
-}
-
-.add-meal-btn {
-    background-color: #34bb19;
-}
-
-h3 {
-    color: #8b8888;
-    margin: 20px 0px;
-    font-size: 22px;
-}
-
-.form-container {
-    display: flex;
-    justify-content: space-evenly;
-    margin-top: 20px;
-}
-
-mat-form-field {
-    width: 450px;
-}
-
-.dropzone {
-    width: 450px;
-    height: 200px;
-    text-align: center;
-    border: none;
-    border-radius: 2px;
-    background-color: #ece7e7;
-    position: relative;
-}
-
-.drop-icon {
-    color: #38424c;
-    font-size: 60px;
-    padding: 10px 30px;
-}
-
-h4 {
-    font-weight: bold;
-    color: #38424c;
-    margin-bottom: 10px;
-}
-
-label {
-    color: #ffffff;
-    border-radius: 10px;
-    background-color: #f50756;
-    padding: 8px 16px;
-}
-
-#upload-img {
-    opacity: 0;
-    position: absolute;
-    z-index: 2;
-    width: 100%;
-    height: 100%;
-    top: 0;
-    left: 0;
-    cursor: pointer;
-}
-
-.inserted-img {
-    width: 450px;
-}
-
-.inserted-img-container,
-.list-container {
-    position: relative;
-}
-
-.delete-image-btn {
-    background-color: #0c5bc23f;
-    margin: 10px 10px 0px 0px;
-    position: absolute;
-    right: 0px;
-    top: 0px;
-}
-
-.add-item-list-btn {
-    background-color: #0c5bc2;
-    position: absolute;
-    margin-top: 10px;
-    left: 47%;
-}
-
-.delete-image-btn:hover,
-.delete-item-list-btn:hover {
-    background-color: #0c5bc2;
-}
-
-.recipe-area {
-    display: block;
-    width: 450px;
-}
-
-.ingredients-list {
-    width: 450px;
-    max-width: 100%;
-    border: solid 1px #8b8888;
-    min-height: 60px;
-    display: block;
-    background: #ece7e7;
-    border-radius: 4px;
-    overflow: hidden;
-}
-
-.ingredient-box {
-    padding: 20px 10px;
-    border-bottom: solid 1px #8b8888;
-    color: rgba(0, 0, 0, 0.87);
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-between;
-    box-sizing: border-box;
-    cursor: move;
-    background: #ece7e7;
-    font-size: 14px;
-}
-
-.delete-item-list-btn {
-    background-color: #0c5bc23f;
-    margin-right: 10px;
-    position: absolute;
-    right: 0px;
-}
-
-.cdk-drag-preview {
-    box-sizing: border-box;
-    border-radius: 4px;
-    box-shadow: 0 5px 5px -3px rgba(0, 0, 0, 0.2), 0 8px 10px 1px rgba(0, 0, 0, 0.14),
-        0 3px 14px 2px rgba(0, 0, 0, 0.12);
-}
-
-.cdk-drag-placeholder {
-    opacity: 0;
-}
-
-.cdk-drag-animating {
-    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
-}
-
-.validation-container {
-    color: #cc0707;
-}
