+++ src/app/components/registration-form/registration-form.component.html
@@ -76,5 +76,5 @@
                      </div>
               </div>
        </div>
-       <button>ZAREJESTRUJ SIĘ</button>
+       <button>Zarejestruj się</button>
 </div>
