+++ src/app/components/registration-form/registration-form.component.scss
@@ -16,6 +16,8 @@
 }
 
 h3 {
+    color: #38424c;
+    font-size: 22px;
     text-align: center;
 }
 
@@ -31,12 +33,12 @@ input {
 
 button {
     width: 202px;
-    height: 42px;
-    background-color: #f01717;
+    padding: 15px;
     border: none;
-    border-radius: 2px;
+    background-color: #1b84b4;
     color: #ffffff;
     font-size: 10px;
     font-weight: bold;
+    border-radius: 2px;
     cursor: pointer;
 }
