+++ src/styles.scss
@@ -45,4 +45,5 @@ body {
 body {
   margin: 0;
   font-family: Roboto, "Helvetica Neue", sans-serif;
+  background: url(C:\Users\Iga\Downloads\andrey-haimin-q2Fyzn-KJOQ-unsplash.jpg);
 }
